+++ src/cli/release/download/download.js
@@ -3,16 +3,13 @@ const fs = require('fs');
 const shell = require('shelljs');
 const chalk = require('chalk');
 const semver = require('semver');
-const { getPlatform, RELEASE_ASSETS, DEFAULT_MIRROR } = require('core/env');
-const { isForgeBinExists, getGlobalForgeVersion } = require('core/forge-fs');
+const { getPlatform } = require('core/env');
+const { isReleaseBinExists, getGlobalForgeVersion } = require('core/forge-fs');
 const { printError, printInfo, printSuccess } = require('core/util');
 const debug = require('core/debug')('download');
-const {
-  fetchAssetInfo,
-  downloadAsset,
-  expandReleaseTarball,
-  fetchReleaseVersion,
-} = require('cli/node/install/install');
+const { downloadAssets, fetchReleaseVersion } = require('cli/node/install/install');
+
+const { DEFAULT_MIRROR, RELEASE_ASSETS } = require('../../../constant');
 
 // eslint-disable-next-line consistent-return
 async function main({ args: [userVersion], opts: { mirror = DEFAULT_MIRROR, releaseDir } }) {
@@ -29,10 +26,12 @@ async function main({ args: [userVersion], opts: { mirror = DEFAULT_MIRROR, rele
     const userVer =
       userVersion && semver.coerce(userVersion) ? semver.coerce(userVersion).version : '';
     const version = userVer || fetchReleaseVersion({ mirror, releaseDir });
+    const unDownloadAssets = RELEASE_ASSETS.filter(x => !isReleaseBinExists(x, version));
     const currentVersion = getGlobalForgeVersion();
-    if (isForgeBinExists(currentVersion)) {
+    if (unDownloadAssets.length === 0) {
       if (version === currentVersion) {
-        return process.exit(1);
+        printInfo(`forge v${version} is already downloaded`);
+        return process.exit(0);
       }
     }
 
@@ -40,13 +39,16 @@ async function main({ args: [userVersion], opts: { mirror = DEFAULT_MIRROR, rele
       printInfo(`Download latest version: v${version}`);
     }
 
-    // Start download and unzip
-    for (const asset of RELEASE_ASSETS) {
-      const assetInfo = fetchAssetInfo({ platform, version, key: asset, mirror, releaseDir });
-      debug(asset, assetInfo);
-      // eslint-disable-next-line no-await-in-loop
-      const assetTarball = await downloadAsset({ asset: assetInfo });
-      expandReleaseTarball(assetTarball, asset, version);
+    const isSuccess = await downloadAssets(unDownloadAssets, {
+      platform,
+      version,
+      mirror,
+      releaseDir,
+    });
+
+    if (!isSuccess) {
+      printError('Please check your assets version or mirror address is correct and try again.');
+      process.exit(1);
     }
 
     printSuccess(`Congratulations! forge v${version} download successfully!`);